+++ Sources/LightPortal/Areas/Area.php
@@ -14,13 +14,13 @@
 
 namespace Bugo\LightPortal\Areas;
 
-use Bugo\LightPortal\Lists\IconList;
-
 if (! defined('SMF'))
 	die('No direct access...');
 
 trait Area
 {
+	use Query;
+
 	public function createBbcEditor(string $content = ''): void
 	{
 		$editorOptions = [
@@ -40,6 +40,10 @@ trait Area
 		$this->checkSubmitOnce('register');
 
 		$this->prepareIconList();
+
+		$this->prepareTopicList();
+
+		$this->prepareMemberList();
 
 		$languages = empty($this->modSettings['userLanguage']) ? [$this->language] : array_unique([$this->context['user']['language'], $this->language]);
 
@@ -125,47 +129,6 @@ trait Area
 		$this->context['lp_num_queries']++;
 	}
 
-	public function prepareIconList(): void
-	{
-		if ($this->request()->hasNot('icons'))
-			return;
-
-		$data = $this->request()->json();
-
-		if (empty($search = $data['search']))
-			return;
-
-		$search = trim($this->smcFunc['strtolower']($search));
-
-		$all_icons = $this->getFaIcons();
-		$template = '<i class="%1$s fa-fw" aria-hidden="true"></i>&nbsp;%1$s';
-
-		$this->hook('prepareIconList', [&$all_icons, &$template]);
-
-		$all_icons = array_filter($all_icons, fn($item) => str_contains($item, $search));
-
-		$results = [];
-		foreach ($all_icons as $icon) {
-			$results[] = [
-				'innerHTML' => sprintf($template, $icon),
-				'value'     => $icon
-			];
-		}
-
-		exit(json_encode($results));
-	}
-
-	public function getFaIcons(): array
-	{
-		if (($icons = $this->cache()->get('all_icons', 30 * 24 * 60 * 60)) === null) {
-			$icons = (new IconList)->getList();
-
-			$this->cache()->put('all_icons', $icons, 30 * 24 * 60 * 60);
-		}
-
-		return $icons;
-	}
-
 	public function getPreviewTitle(string $prefix = ''): string
 	{
 		return $this->getFloatSpan(
@@ -177,7 +140,7 @@ trait Area
 		) . '<br>';
 	}
 
-	private function getFloatSpan(string $text, string $direction = 'left'): string
+	public function getFloatSpan(string $text, string $direction = 'left'): string
 	{
 		return '<span class="float' . $direction . '">' . $text . '</span>';
 	}
