--- Themes/default/LightPortal/partials/keywords.php
@@ -1,24 +1,0 @@
-<?php
-
-global $context, $txt, $modSettings;
-
-echo '
-	<script>
-		VirtualSelect.init({
-			ele: "#keywords",', ($context['right_to_left'] ? '
-			textDirection: "rtl",' : ''), '
-			dropboxWrapper: "body",
-			maxWidth: "100%",
-			multiple: true,
-			search: true,
-			markSearchResults: true,
-			showValueAsTags: true,
-			allowNewOption: true,
-			showSelectedOptionsFirst: true,
-			placeholder: "', $txt['lp_page_keywords_placeholder'], '",
-			noSearchResultsText: "', $txt['no_matches'], '",
-			searchPlaceholderText: "', $txt['search'], '",
-			clearButtonText: "', $txt['remove'], '",
-			maxValues: ', $modSettings['lp_page_maximum_keywords'] ?? 10, '
-		});
-	</script>';
